Pipeing through packet and bytes ctr vals

+++ src/libbpfilter/bpfilter.h
@@ -22,7 +22,7 @@ struct nlmsghdr;
  */
 int bf_cli_set_chain(const struct bf_chain *chain);
 
-int bf_cli_get_ctrs(const char *chain_name, const char *rule_name, int *ctrs);
+int bf_cli_get_ctrs(const char *chain_name, const char *rule_name, uint64_t *ctrs);
 
 /**
  * Send iptable's ipt_replace data to bpfilter daemon.
